+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [#148](https://github.com/smile-io/ember-polaris/pull/148) [FEATURE] Add `polaris-label` and `polaris-labelled` components.
 - [#151](https://github.com/smile-io/ember-polaris/pull/151) [FIX] Add missing `hidden` class to `polaris-label`.
 - [#152](https://github.com/smile-io/ember-polaris/pull/152) Accept card title as a string or a component
+- [#154](https://github.com/smile-io/ember-polaris/pull/154) [FEATURE] Add `polaris-range-slider` component.
 
 ### v1.6.1 (June 28, 2018)